refactor(rules-of-hooks): inline code-path-analysis modules
Merge code-path-analysis directory contents into rules-of-hooks.ts:
- Inline code-path-analyzer, code-path-segment, code-path-state
- Inline code-path, fork-context, id-generator, assert utilities
- Remove separate code-path-analysis directory and LICENSE
- Update eslint.config.ts and tsl.config.ts to ignore ported rules
