--- packages/plugins/eslint-plugin-react-x/src/utils/code-path-analysis/README.md
@@ -1,6 +1,0 @@
-# Code Path Analyzer
-
-This code is a forked version of ESLints Code Path Analyzer which includes
-support for Component Syntax.
-
-Forked from: https://github.com/eslint/eslint/tree/main/lib/linter/code-path-analysis
