--- packages/plugins/eslint-plugin-react-x/src/utils/code-path-analysis/assert.js
@@ -1,9 +1,0 @@
-"use strict";
-
-function assert(cond) {
-  if (!cond) {
-    throw new Error("Assertion violated.");
-  }
-}
-
-export default assert;
