+++ packages/plugins/eslint-plugin-react-x/tsl.config.ts
@@ -14,7 +14,8 @@ export default defineConfig({
         "**/*.d.ts",
         "**/dist/**",
         "**/build/**",
-        "src/utils/code-path-analysis",
+        "src/rules/exhaustive-deps.ts",
+        "src/rules/rules-of-hooks.ts",
       ],
       { ignore: ["**/node_modules/**"] },
     ),
